requirements.txt
streamlit
openai
pandas
snowflake-connector-python
